LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-headers.tentative.https.sub.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

PASS Sec-Fetch-Storage-Access is omitted when credentials are omitted
FAIL Sec-Fetch-Storage-Access is `none` when unpartitioned cookies are unavailable. assert_array_equals: value is undefined, expected array
TIMEOUT Sec-Fetch-Storage-Access is `inactive` when unpartitioned cookies are available but not in use. Test timed out
NOTRUN Sec-Fetch-Storage-Access is `active` after a valid retry with matching explicit allowed-origin.
NOTRUN Sec-Fetch-Storage-Access is active after retry with wildcard `allowed-origin` value.
NOTRUN 'Activate-Storage-Access: retry' is a no-op on a request without an `allowed-origin` value.
NOTRUN 'Activate-Storage-Access: retry' is a no-op on a request from an origin that does not match its `allowed-origin` value.
NOTRUN Activate-Storage-Access `retry` is a no-op on a request with a `none` Storage Access status.
NOTRUN Activate-Storage-Access `load` header grants storage access to frame.
NOTRUN Activate-Storage-Access `load` is honored for `active` cases.
NOTRUN Activate-Storage-Access `load` header is a no-op for requests without storage access.
NOTRUN Sec-Fetch-Storage-Access is `inactive` for ABA case.
NOTRUN Storage Access can be activated for ABA cases by retrying.
NOTRUN Sec-Fetch-Storage-Access maintains value on same-origin redirect.
NOTRUN Sec-Fetch-Storage-Access is not 'active' after cross-origin same-site redirection.
NOTRUN Sec-Fetch-Storage-Access loses value on a cross-site redirection.
NOTRUN Activate-Storage-Access retry is handled before any redirects are followed.
